load and display arbitrary data (FAILED)
the first thing I tried to do was display raw data as an image. However this turns out to be not as simple as I thought. Disabling the header check isn't enough to get the image to finish processing. The header information is needed to actually read the image. Kind of a duh moment! So for now I'm going to put aside the goal of loading ANY image and just focus on image manipulation.