port custom ui code
was about to use raygui, but I was having trouble figuring out how to add it to the project, so I figured it would be easier if I ported in the UI code that I use in my game engine. Removed lots of specific functionality that was specific to that engine. Now we can build up functionlity for this app!