Ui iteration, commands
design realization: in a painting app we have a brush shape and a color. In this app, the brush shape is the same as usual, however the color is the effect that the color has on the image. This means that we so far have 2 brush effects and 1 brush shape.