scaling memory based on image
a game will know it's memory requirements and commit everything at the start, but I think we benifit from doing some scaling of memory based on the size of our image. This way our memory footprint is exactly what it needs to be based on what we're editing, and we can release that memory if we start working on a smaller image.