switch to conditional painting
paint is only added when it's adding something new to the data. this is more expensive in now we're doing a conditional on each pixel, but I don't think it's that big of a deal. our bottleneck is now uploading the final image so frequently because we're so fast now!