realize compression and decoding is unnecessary!!!
I feel a little silly for only JUST not realizing this. I was running under the assumption that these corruptions were due to compression and decompression. but no, it's actually only relevant to defiltering. The filtering is to make compression better, but for our purposes we can just skip compression and decompression altogether since we're only corrupting the filtered image! This makes even large images not instant, but incredibly faster! And now most averaged sized images are basically instant. wowzers!