remove unneded compression related code & fpng dependency
there are perhaps interesting things to be explored with png corruption in how it is compressed and uncompressed, but for now I'm going to remove these. Thanks to version control I can always come find it and put it back if I need it. Thank you for your service fpng you have been amazing o7