update canvas visualization, colors map directly to filtered colors
while researching i found a visualization of the filtered png that made me realize, you can just interpret the filtered data as colors and you can get a canvas that's the same size as your image. I'm undecided if i like this better. The previous version mapped every byte to a color intensity. this is doing some interpretation of that data and also hiding some from you (the filtered bit) and also changes the transformation you're doing a good deal, as now 1 pixel affects 4 bytes at once. but perhaps that's more intuitive than dealing with this stretched out thing that only kind of resembles your image...I'll switch to this for now, I think it's better, but later you should be able to customize this setup heavily