ui proper positioning, layout, & alignemnt
time will tell if this iteration of layout works, but i thought of something cool that we're gonna try. in previous UI systems that i've worked on, the alignment of a block (if it's left, center, or right aligned) is set by the parent block. We still use the parent block for the layout (if it goes top to bottom or left to right), but each block now controls its own alignmnet within the parent. positioning is determined by only blocks that match alignment type allowing blocks of differnet alignment to exists in the same block. Now we can trivally set things to the left or right. might be the wrong move to have blocks of differnt layout logics in the same block but we'll see how it goes