start sketching out new ui, ui bug fixes, weird ui line drawing
after actually trying to build ui with this new idea I'm convinvced it's not it chief. its proving to be rather difficult to track where block should go, especially since this logic is independent to each axis. Alignment for both start center and end can be better implemented via a good fill type. this also forces the ui blocks to have a story for how they interact with each other