test tool, swap pixels in the test tool, show on image break
thinking on how we ought to add more image manipulation features (and trying it directly). reading up on glitch techniques and ideology. this tool where you have so much control over the glitches is very antithetical to the philosophy of glitch. but unsure which came first, the philosophy, or the process, and which is more important to perserve here...