experimental hack: paint root image
am realizing - we kind of want to paint either or the original image, the filtered data, or the final image. but what does that mean? Realizing out painting needs to be all painting data transformations. we can't edit pixels for sources that are updated continuously like the filtered data. if all painting is data transformations, we can paint anywhere!