rename sum of children to fit children
this size label that i've used for a long time is wrong. its technically a sum or max of children. but also with this new interaction with fill the true definitoin is now much more nuanced. time will tell if this implementation is too complex to be intuitive to use